+++ src/steps/index.test.ts
@@ -7,10 +7,17 @@ import { IntegrationConfig, sanitizeConfig } from '../config';
 import { fetchMembers } from './members';
 import { fetchRepos } from './repos';
 import { fetchTeams } from './teams';
+import { fetchCollaborators } from './collaborators';
 import { fetchPrs } from './pullrequests';
 import { fetchAccountDetails } from './account';
+import {
+  GITHUB_COLLABORATOR_ENTITY_TYPE,
+  GITHUB_REPO_USER_RELATIONSHIP_TYPE,
+} from '../constants';
 import { integrationConfig } from '../../test/config';
 import { setupGithubRecording } from '../../test/recording';
+
+jest.setTimeout(10000);
 
 let recording: Recording;
 afterEach(async () => {
@@ -43,6 +50,7 @@ test('should collect data', async () => {
   await fetchMembers(context);
   await fetchRepos(context);
   await fetchTeams(context);
+  await fetchCollaborators(context);
   await fetchPrs(context);
 
   // Review snapshot, failure is a regression
@@ -139,7 +147,6 @@ test('should collect data', async () => {
     },
   });
 
-  /*
   const prs = context.jobState.collectedEntities.filter((e) =>
     e._class.includes('PR'),
   );
@@ -159,5 +166,17 @@ test('should collect data', async () => {
       },
       required: ['webLink', 'displayName'],
     },
-  });*/
+  });
+
+  const repoUserRelationships = context.jobState.collectedRelationships.filter(
+    (r) => r._type === GITHUB_REPO_USER_RELATIONSHIP_TYPE,
+  );
+  expect(repoUserRelationships.length).toBeGreaterThan(0);
+
+  const outsideCollaboratorEntities = context.jobState.collectedEntities.filter(
+    (e) =>
+      e._type === GITHUB_COLLABORATOR_ENTITY_TYPE &&
+      e.role === 'outside collaborator',
+  );
+  expect(outsideCollaboratorEntities.length).toBeGreaterThan(0);
 });