+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './integration';
 export * from './client';
 export * from './entities';
+export * from './relationships';